Update hatchling
Make the shadows under the moving blocks in hatchling turn red when the block is over the trash. Make the note block only play a preview sound when changing the note, not when updating duration. Add the option for easing to move animations. Update colors for keyboard in note block popup. Update hatchling level switching animation with easing and to fix bugs. Give the hatchling ABC block two rows of buttons. Remove the stopwatch icon from the hatchling sound block popup. Add moving degree diagram to hatchling servo block popup. Adjust size and position of pointer triangle in hatchling bubble overlays.

+++ UIParts/HLLevelSwitch.js
@@ -89,7 +89,7 @@ HLLevelSwitch.prototype.press = function() {
 	this.animations[0] = GuiElements.animate.updateColor(tB, Colors.white, duration)
 	this.animations[1] = GuiElements.animate.updateColor(tW, this.textColor, duration)
 	this.animations[2] = GuiElements.animate.move(this.circleE, x, 0, duration)
-	this.animations[3] = GuiElements.animate.move(this.tempG, oldTabX, 0, duration)
+	this.animations[3] = GuiElements.animate.move(this.tempG, oldTabX, 0, duration, true)
 
 	
 	setTimeout(function() {
@@ -114,15 +114,18 @@ HLLevelSwitch.prototype.setSwitch = function(level) {
 		this.animations[i].remove()
 	}
 
-	this.oldTab.dontDelete = false
-	this.oldTab.delete()
+	if (this.oldTab != null) {
+		this.oldTab.dontDelete = false
+		this.oldTab.delete()
+	}
+	
 	TabManager.activeTab.mainG.removeAttributeNS(null, "clip-path")
 	GuiElements.move.group(TabManager.activeTab.mainG, 0, 0)
 	GuiElements.layers.activeTab.appendChild(TabManager.activeTab.mainG)
-	this.tempG.remove()
-	this.tempG = null
-	console.log(TabManager.activeTab.mainG)
-	
+	if (this.tempG != null) {
+		this.tempG.remove()
+		this.tempG = null
+	}
 
 	switch(level){
 	case 1:
